Round1B_Persona_Analysis/requirements.txt
pymupdf==1.24.9
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1
pyyaml==6.0.2
networkx==3.3
rapidfuzz==3.9.7
